[140] Fix compatibility with rev 4577.
Fix compatibility with rev 4577.
Fixed crash in Curator Script. 

git-svn-id: https://scriptdev2.svn.sourceforge.net/svnroot/scriptdev2@140 5f9c896b-1e26-0410-94da-f77f675e2462

+++ scripts/zone/gruuls_lair/instance_gruuls_lair.cpp
@@ -113,7 +113,7 @@ struct MANGOS_DLL_DECL instance_gruuls_lair : public ScriptedInstance
         }
         else if(type == "Event_Maulgar_Started_data2")
         {
-            Maulgar_Event_Starter += data << 32;
+            Maulgar_Event_Starter += (uint64(data) << 32);
 
         }else if(type == "Event_Gruul_Started")
             GruulInProgress = true;
